Tidied up Go Files
